packages/fastapi-base-template-repository/pyproject.toml
[tool.poetry]
name = "fastapi-base-template-repository"
version = "0.1.0"
description = ""
authors = ["Paras <paras.bhardava@jeavio.com>"]
readme = "README.md"
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
fastapi = ">=0.115.8,<0.116.0"
uvicorn = {extras = ["standard"], version = ">=0.34.0,<0.35.0"}
gunicorn = ">=23.0.0,<24.0.0"
uvloop = ">=0.21.0,<0.22.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.41"}
pydantic-settings = "^2.0.0"
yarl = "^1.9.0"
pydantic = "^2.11.5"
loguru = "^0.7.3"
asyncpg = "^0.30.0"
psycopg = {extras = ["c"], version = "^3.2.4"}
alembic = "^1.16.1"
pyyaml = "^6.0.2"
sqlmodel = "^0.0.24"
redis = "^6.2.0"
slowapi = "^0.1.9"
# OpenTelemetry dependencies
opentelemetry-api = "^1.21.0"
opentelemetry-sdk = "^1.21.0"
opentelemetry-exporter-otlp-proto-grpc = "^1.21.0"
opentelemetry-instrumentation-fastapi = "^0.42b0"
opentelemetry-instrumentation-sqlalchemy = "^0.42b0"
opentelemetry-instrumentation-logging = "^0.42b0"
opentelemetry-instrumentation-requests = "^0.42b0"
opentelemetry-exporter-otlp = "^1.21.0"
opentelemetry-instrumentation-urllib3 = "^0.42b0"
opentelemetry-instrumentation-redis = "^0.42b0"
# Sentry
sentry-sdk = "^1.44.1"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
pylint = "^3.3.7"
mypy = "^1.16.0"
ruff = "^0.11.13"
types-pyyaml = "^6.0.12.20250516"
pytest = "^8.4.0"
